s3 image upload

+++ app/src/main/java/com/example/finalapp/model/DirectChat.kt
@@ -2,6 +2,6 @@ package com.example.finalapp.model
 
 data class DirectChat(
     val userId:String,
-    val lat:Float,
-    val long:Float,
+    val lat: Double,
+    val long: Double,
 )
